Cleaning up validation
